% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smlsom-functions.R
\name{predict.smlsom_clf.mvn}
\alias{predict.smlsom_clf.mvn}
\title{prediction function for smlsom_clf.mvn}
\usage{
\method{predict}{smlsom_clf.mvn}(object, newdata, type = c("class", "probability"))
}
\arguments{
\item{object}{smlsom_clf.mvn object.}

\item{newdata}{a numeric (n, p) matrix.}

\item{type}{type of return. \code{class} gives a vector of length n whose elements are the classes with the largest posterior probability. \code{probability} provides the posterior probabilities matrix.}
}
\value{
hogehoge
}
\description{
This function provides smlsom clustering.
}
