% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smlsom-functions.R
\name{smlsom_clf.mvn}
\alias{smlsom_clf.mvn}
\title{smlsom for classification problem using multivariate normals}
\usage{
smlsom_clf.mvn(
  X,
  y,
  K,
  Mk = 5,
  beta = 5,
  niter = nrow(data),
  Mus,
  Sigmas,
  alpha = c(0.05, 0.01),
  radii
)
}
\arguments{
\item{X}{A numeric (n, p) matrix.}

\item{y}{class labels. n integer vector.}

\item{K}{The number of classes.}

\item{Mk}{The number of subclasses within each class.}

\item{beta}{hogehoge}

\item{niter}{The number of iteration. Default is \code{nrow(data)}.}

\item{Mus}{A numeric (M, p) matrix. The rows are the mean vector of each cluster. Default is to cluster \code{data} randomly.}

\item{Sigmas}{A numeric (p, p, M) array. \code{Sigmas[,,m]} is the variance-covariance matrix of the mth cluster. Default is identity matrices.}

\item{alpha}{Range of learning rates: \eqn{\alpha = (\alpha_1, \alpha_2)^t}. Note that \eqn{\alpha_1 > \alpha_2 > 0}. Monotonically decreasing from \eqn{\alpha_1} to \eqn{\alpha_2} for the number of iterations. Default is \code{alpha=c(0.05, 0.01)}. See \code{kohonen} package for details.}

\item{radii}{Range of neighbourhood radius: \eqn{r = (r_1, r_2)^t}. Note that \eqn{r_1 > r_2 > 0}. Monotonically decreasing from \eqn{r_1} to \eqn{r_2} for the number of iterations. Default is to start with a value that covers 2/3 of node distances. See \code{kohonen} package for detals.}
}
\value{
hogehoge
}
\description{
This function provides smlsom clustering.
}
